# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "El Mexicano API",
  "description": "Backend réplica (demo) de **Transportes El Mexicano S.R.L.**: venta de pasajes de bus, rutas e itinerarios, carga y encomiendas, puerta a puerta y centro de ayuda.\n\nFechas en hora de Bolivia (America/La_Paz) y montos en bolivianos (Bs).",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/oficinas": {
   "get": {
    "tags": [
     "Catálogos"
    ],
    "summary": "Boleterías y bodegas con horario",
    "operationId": "oficinas_api_v1_oficinas_get",
    "parameters": [
     {
      "name": "ciudad",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "examples": [
        "Santa Cruz"
       ],
       "title": "Ciudad"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/OficinaOut"
         },
         "title": "Response Oficinas Api V1 Oficinas Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/feriados": {
   "get": {
    "tags": [
     "Catálogos"
    ],
    "summary": "Feriados nacionales y departamentales",
    "operationId": "feriados_api_v1_feriados_get",
    "parameters": [
     {
      "name": "desde",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "description": "Por defecto, hoy",
       "title": "Desde"
      },
      "description": "Por defecto, hoy"
     },
     {
      "name": "hasta",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "description": "Por defecto, 90 días después de «desde»",
       "title": "Hasta"
      },
      "description": "Por defecto, 90 días después de «desde»"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/FeriadoOut"
         },
         "title": "Response Feriados Api V1 Feriados Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/faqs": {
   "get": {
    "tags": [
     "Centro de ayuda"
    ],
    "summary": "Preguntas frecuentes por categoría",
    "operationId": "faqs_api_v1_faqs_get",
    "parameters": [
     {
      "name": "categoria",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "examples": [
        "pasajeros"
       ],
       "title": "Categoria"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/FaqCategoriaOut"
         },
         "title": "Response Faqs Api V1 Faqs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/paginas/{slug}": {
   "get": {
    "tags": [
     "Centro de ayuda"
    ],
    "summary": "Una página de contenido",
    "operationId": "pagina_api_v1_paginas__slug__get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PaginaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/oficinas": {
   "get": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Listar oficinas",
    "operationId": "listar_api_v1_admin_oficinas_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "description": "Cantidad de resultados",
       "default": 20,
       "title": "Limit"
      },
      "description": "Cantidad de resultados"
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Resultados a saltar",
       "default": 0,
       "title": "Offset"
      },
      "description": "Resultados a saltar"
     },
     {
      "name": "ciudad_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por ciudad_id"
     },
     {
      "name": "tipo",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por tipo"
     },
     {
      "name": "activo",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por activo"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Pagina_OficinaAdminOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Crear oficinas",
    "operationId": "crear_api_v1_admin_oficinas_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OficinaIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OficinaAdminOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/vehiculos-carga": {
   "get": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Listar vehículos de carga",
    "operationId": "listar_api_v1_admin_vehiculos_carga_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "description": "Cantidad de resultados",
       "default": 20,
       "title": "Limit"
      },
      "description": "Cantidad de resultados"
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Resultados a saltar",
       "default": 0,
       "title": "Offset"
      },
      "description": "Resultados a saltar"
     },
     {
      "name": "tipo",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por tipo"
     },
     {
      "name": "ciudad_base_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por ciudad_base_id"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Pagina_VehiculoOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Crear vehículos de carga",
    "operationId": "crear_api_v1_admin_vehiculos_carga_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VehiculoIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VehiculoOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/rutas": {
   "get": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Listar rutas",
    "operationId": "listar_api_v1_admin_rutas_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "description": "Cantidad de resultados",
       "default": 20,
       "title": "Limit"
      },
      "description": "Cantidad de resultados"
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Resultados a saltar",
       "default": 0,
       "title": "Offset"
      },
      "description": "Resultados a saltar"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Pagina_RutaAdminOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/plantillas-horario": {
   "get": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Listar plantillas de horario",
    "operationId": "listar_api_v1_admin_plantillas_horario_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "description": "Cantidad de resultados",
       "default": 20,
       "title": "Limit"
      },
      "description": "Cantidad de resultados"
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Resultados a saltar",
       "default": 0,
       "title": "Offset"
      },
      "description": "Resultados a saltar"
     },
     {
      "name": "ruta_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por ruta_id"
     },
     {
      "name": "activo",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por activo"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Pagina_PlantillaOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Crear plantillas de horario",
    "operationId": "crear_api_v1_admin_plantillas_horario_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PlantillaIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlantillaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/tarifas-pasaje": {
   "get": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Listar tarifas de pasaje",
    "operationId": "listar_api_v1_admin_tarifas_pasaje_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "description": "Cantidad de resultados",
       "default": 20,
       "title": "Limit"
      },
      "description": "Cantidad de resultados"
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Resultados a saltar",
       "default": 0,
       "title": "Offset"
      },
      "description": "Resultados a saltar"
     },
     {
      "name": "ruta_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por ruta_id"
     },
     {
      "name": "tipo_asiento_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por tipo_asiento_id"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Pagina_TarifaPasajeOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Crear tarifas de pasaje",
    "operationId": "crear_api_v1_admin_tarifas_pasaje_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TarifaPasajeIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TarifaPasajeOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/tarifas-carga": {
   "get": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Listar tarifas de carga",
    "operationId": "listar_api_v1_admin_tarifas_carga_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "description": "Cantidad de resultados",
       "default": 20,
       "title": "Limit"
      },
      "description": "Cantidad de resultados"
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Resultados a saltar",
       "default": 0,
       "title": "Offset"
      },
      "description": "Resultados a saltar"
     },
     {
      "name": "origen_ciudad_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por origen_ciudad_id"
     },
     {
      "name": "destino_ciudad_id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por destino_ciudad_id"
     },
     {
      "name": "tipo_envio",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      },
      "description": "Filtrar por tipo_envio"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Pagina_TarifaCargaOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Crear tarifas de carga",
    "operationId": "crear_api_v1_admin_tarifas_carga_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TarifaCargaIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TarifaCargaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/feriados/{feriado_id}": {
   "delete": {
    "tags": [
     "Admin · Catálogos"
    ],
    "summary": "Eliminar feriado",
    "operationId": "eliminar_feriado_api_v1_admin_feriados__feriado_id__delete",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "feriado_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Feriado Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "FeriadoOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "fecha": {
      "type": "string",
      "format": "date",
      "title": "Fecha"
     },
     "nombre": {
      "type": "string",
      "title": "Nombre"
     },
     "departamento": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Departamento"
     }
    },
    "type": "object",
    "required": [
     "id",
     "fecha",
     "nombre",
     "departamento"
    ],
    "title": "FeriadoOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Pagina_VehiculoOut_": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/VehiculoOut"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "total",
     "limit",
     "offset",
     "items"
    ],
    "title": "Pagina[VehiculoOut]"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    }
   }
  }
 }
}